incorrect loading of last page

+++ src/main/groovy/com/riversoft/eventsion/display/models/FirmwareSendModel.groovy
@@ -1,15 +1,13 @@
 package com.riversoft.eventsion.display.models
 
-/**
- * Created by gor on 30.11.16.
- */
+
 class FirmwareSendModel {
 
     Byte last
 
     Short number
 
-    Byte length
+    Short length
 
     byte[] payload
 }
